Refactor parsers.util into modular package structure
Refactor the monolithic goedels_poetry/parsers/util.py (3499 lines) into
a structured package with 26 focused modules organized into 5 logical
subdirectories. This improves code organization, maintainability, and
readability while preserving full backward compatibility.

Package Structure:
- foundation/: Low-level utilities, constants, and basic AST operations
  * constants.py: Type aliases and constants
  * kind_utils.py: AST kind normalization and checking
  * ast_validation.py: AST structure validation
  * nested_extraction.py: Safe nested dictionary/list access
  * ast_walkers.py: Generic AST traversal functions
  * goal_context.py: Goal context parsing
  * ast_to_code.py: AST to code serialization

- names_and_bindings/: Name and binding extraction
  * name_extraction.py: Extract names from declarations and have statements
  * anonymous_haves.py: Anonymous have statement collection and naming
  * binding_name_extraction.py: Extract names from binding types
  * binding_value_extraction.py: Extract values from binding types

- types_and_binders/: Type extraction and binder construction
  * type_extraction.py: Extract type ASTs from various node types
  * binder_construction.py: Construct binder AST nodes
  * binding_types.py: Determine types for bindings

- collection_and_analysis/: Declaration collection and analysis
  * decl_collection.py: Collect named declarations and dependencies
  * reference_checking.py: Check references and containment in AST
  * theorem_binders.py: Extract theorem binders and find earlier bindings

- high_level/: High-level operations
  * sorry_detection.py: Sorry hole detection and tracking
  * named_subgoals.py: Named subgoal AST operations
  * subgoal_rewriting.py: Main subgoal rewriting logic

Changes:
- Split util.py into 20 focused modules (plus 6 __init__.py files)
- Each module is logically coherent and ideally under 500 lines
- All modules use relative imports within the package
- Maintain full backward compatibility via __init__.py re-exports
- Public API functions (_*) remain accessible from goedels_poetry.parsers.util
- Test-required private functions (__*) are also re-exported for compatibility
- All existing code continues to work without modification

Technical Details:
- Fixed missing imports (logging, typing.Any, etc.)
- Fixed import ordering and formatting
- All type checks pass (mypy)
- All linting checks pass (ruff)
- No algorithm changes - functionality is identical

This refactoring maintains 100% backward compatibility while significantly
improving code organization and maintainability.
